test: specify profiles on test DhisConfigration components
