+++ dhis-2/dhis-support/dhis-support-test/src/main/java/org/hisp/dhis/config/PostgresDhisConfiguration.java
@@ -30,12 +30,14 @@ package org.hisp.dhis.config;
 import org.hisp.dhis.external.conf.DhisConfigurationProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 /**
  * Use this Spring configuration for tests relying on the Postgres DB running in a Docker container.
  *
  * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
+@Profile("test-postgres")
 @Configuration
 public class PostgresDhisConfiguration {
   @Bean
